add inheritance layer to json output classes
