+++ lib/wrf_forecast/json_converter.rb
@@ -2,6 +2,7 @@ module WrfForecast
 
   module JsonConverter
     require 'wrf_library/json_converter'
+    require 'wrf_forecast/json_converter/forecast_station_converter'
     require 'wrf_forecast/json_converter/forecast_converter'
   end
